changed sidebar
